Create a new sample
Create the folder SAMPLE.

The "+sys.path.append('../') # Just to make the GIT source more cleaner
without creating a module. In YOUR project you DON'T have to do that"
Is just here to make everything working right out of the box when you
clone the project. In the realife you don't have to do that.